Сделаны правки после ревью

+++ src/main/java/faang/school/accountservice/model/Balance.java
@@ -36,10 +36,10 @@ public class Balance {
     private Account accountNumber;
 
     @Column(name = "current_authorization_balance")
-    private long currentAuthorizationBalance;
+    private double currentAuthorizationBalance;
 
     @Column(name = "current_actual_balance")
-    private long currentActualBalance;
+    private double currentActualBalance;
 
     @CreationTimestamp
     @Temporal(TemporalType.TIMESTAMP)
